Fix playRound logic
The sequence was messed up. The values of the
getHumanChoice and getComputerChoice functions
were not updated at the beginning of the round.
Therefore, the logic of the game were broken.

Now sequence is right. The computer and the person
enter the options, the options are compared and
the result of the round is shown. After that,
the next round begins with the same sequence.